Rename: 컴포넌트 디렉토리 경로 변경
- 컴포넌트 페이지 단위로 그룹핑

#8
