+++ src/components/landing/LandingSlide.ts
@@ -1,6 +1,6 @@
 import { html, css, CSSResultGroup } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
-import { TaingElement } from './Taing';
+import { TaingElement } from '../Taing';
 
 @customElement('landing-slide')
 export class Slide extends TaingElement {
